+++ src/precompile.jl
@@ -4,6 +4,9 @@ let
     to = TimerOutput()
     @timeit to "1" string(1)
     sprint(show, to)
+    cto = ConcurrentTimerOutput()
+    @timeit cto "1" string(1)
+    sprint(show, cto)
 end
 
 function _precompile_()
@@ -17,5 +20,10 @@ function _precompile_()
     @assert Base.precompile(Tuple{typeof(enable_timer!), TimerOutput})
     @assert Base.precompile(Tuple{typeof(complement!), TimerOutput})
     @assert Base.precompile(Tuple{typeof(do_accumulate!), Section, UInt64, Int64})
+    @assert Base.precompile(Tuple{typeof(push!), ConcurrentTimerOutput, String})
+    @assert Base.precompile(Tuple{typeof(pop!), ConcurrentTimerOutput})
+    @assert Base.precompile(Tuple{typeof(register_task_timer!), ConcurrentTimerOutput})
+    @assert Base.precompile(Tuple{typeof(merged), ConcurrentTimerOutput})
+    @assert Base.precompile(Tuple{typeof(reset_timer!), ConcurrentTimerOutput})
     return @assert Base.precompile(Tuple{Type{TimerOutput}, String})
 end